Extending confirmation email timeout
When a user signs up for the first time an email is sent which they
need to open and then click a link to confirm it is correct and
actually theirs. We currently allow them 6 hours to complete this task
but following feedback we would like to extend this to 12.