Further attempt to block "free edits".
We do this by (i) storing the updated Ohm Registration ID in the session after adding the order the to the Registration.   This prevents the user from navigating backwards to the confirmation page using the browser history, to get a free edit; and (ii) clearing out fields related to registration-editing from the session in the "contact us to complete payment" page, to provide an additional layer of safety.