Fix wrong use of country code from countries gem
https://eaflood.atlassian.net/browse/WC-140

This relates to a call that was logged with our ESG team (ESG38945). The issue reported was that attempts to access Worldpay to pay for a registration would fail for some users.

Specifically the error being returned by Worldpay was `Invalid address: country code 'LO' is not a valid ISO-3166 country code.` After some research we identified **LO** is the code for Slovakia, so assume a user had entered there address and specifed the country as Slovakia.

The country code is a required field when dealing with worldpay; we have to include one in the XML we pass across when redirecting to it.

After some investigation though we spotted the problem is our incorrect usage of the countries gem. We should be calling `country.alpha2` to get the official ISO3166 code (SK). We were calling `country.gec` which returns **LO**.

GEC refers to the U.S. Standard GEC (Global Embargoed Countries) list, whose codes actually differ in some cases hence we are causing worldpay to error by using it.