Fix broken comp. house strike off test
When running the rspec suite we started getting a couple of failing tests. This is because the company referred to in `companies_house_caller_spec.rb` to test receiving a response that has the sub-status of 'proposal to strike off' is now no longer active.

As the VCR cassettes are now no longer committed to source control (a decision made in 2015 7e72d8c021075d8e61fc7967c76f04f74d7729a9) this means they are generated either when cloning the project for the first time, or updated once a year. As such with the change in status of the previous company used it meant the cassette generated now does not match the expectations of the test.

The only fix was to find another company and update the registration number used in the test.