Refactor existing_registration_ctrl logic
https://eaflood.atlassian.net/browse/WC-84

In anticipation of users who first registered with the digital service in March 2015, we need to update the functionality at the start of the renewals page to better support them.

Initially was going to focus on the content and validations errors in that page to, but upon starting work and reviewing what's there its become more an exercise in housekeeping and refactoring prior to making changes in that area.

Therefore this PR focuses mainly on getting the code ready to start implementing new functionality.

The changes in summary are