Fix mistaken use of match() method
I had incorrectly understood how `String.match()` works. I assumed it was returning true or false as to whether the string matched the regex provided, but what its actually doing is returning from the string that which matches the regex provided. So in our case if the registration number matched the regex, we were getting the registration number back again, else nil.

So with unit tests checking for truthy and falsey values this is fine, but when I started to try and use it within actual code expecting a true or false value I was getting errors.

This change updates the tests to ensure we get a true or false return value from the helpers, and the helpers themselves have been refactored to take the result from `match()` and convert it.