Handle shorter IR number format
When initially looking to refactor the method that determines whether a string matches the IR format, I believed that the format was pretty fixed based on the records I looked at.

However when starting to look at refactoring the `ExistingRegistrationController.create()` method I found its unit tests specified 'CB/AE888XX/A001' as being valid. The difference is that there are only 3 digits in the middle section instead of 4.

Therefore till I've confirmed whether this is actually the case, I've amended the unit tests and `RegistrationsHelper.valid_ir_format?` method to handle both the 15 and 16 character formats.