Remove redundant `valid?` check in create()
Seeing that the behaviour of the controller is to return the user to the same page with a 400 error in the event of the registration being invalid, I started to create a unit test to cover this. Then as I began to refactor the method I'd be confident that's its current behaviour is captured in our tests.

However I then researched what it would take for a registration to be invalid at this point in the journey (`enterRegNumber` step) and found its only if it's missing a reg_uuid. The problem is `setup_registration()` in the registrations helper already checks for this and will throw an error if a registration is passed in with a missing reg_uuid. And `setup_registration()` is one of the first things called in `create()` so the subsequent call to `@registration.valid?` is pointless in this particular case.

This change removes the check to reduce the number of conditional branches in the `create()` method.