Remove helper method formatIRRenewalNumber()
It was a method that simply strips then uppercases whatever string is passed in, and this can be done as `foo.strip.upcase`.

The method is therefore unnecessary, has no unit tests, and is currently only used in one place anyway.

This change removes it and updates the code that previously used it to do the work in place.