Refactor existing IR logic in existing_reg_ctrlr
We will be adding new logic to handle new style registrations, which in turn will make the `ExistingRegistrationController` grow.

So prior to that this change refactors the code so the logic for determining if the entered registration number is a valid IR registration is moved to a new method.

This will help reduce complexity in the controller method and make its behaviour clearer.