Refactor existing_registration_ctrl logic (#111)
https://eaflood.atlassian.net/browse/WC-84

In anticipation of users who first registered with the digital service in March 2015, we need to update the functionality at the start of the renewals page to better support them.

Initially was going to focus on the content and validations errors in that page to, but upon starting work and reviewing what's there its become more an exercise in housekeeping and refactoring prior to making changes in that area.

Therefore this PR focuses mainly on getting the code ready to start implementing new functionality.

The changes in summary are

* Updated example to include new style reg. number

Include an example of an upper tier registration number in the helper text for the page. Also rename the locale key to be more generic.

* Rename isCurrentRegistrationType to match ruby style

Also added missing unit tests for this method and refactored the internals to match the logic which will be used in the new renewals service.

It also means we simplify the internal logic.

Its now called `valid_registration_format?`.

* Rename isIRRegistrationType to match ruby style

Also added missing unit tests for this method and refactored the internals to match the logic which will be used in the new renewals service.

It also means we simplify the internal logic.

Its now called `valid_ir_format?`.

Also improved the regular expression used to improve the matching against the entered IR number. This should increase the likely that

- we reduce the number of unnecessary calls to the java services using invalid IR references
- we better report the issue with the reference to the user; we didn't find a matching reg. because the number provided is invalid

* Remove redundant `valid?` check in create()

Seeing that the behaviour of the controller is to return the user to the same page with a 400 error in the event of the registration being invalid, I started to create a unit test to cover this. Then as I began to refactor the method I'd be confident that's its current behaviour is captured in our tests.

However I then researched what it would take for a registration to be invalid at this point in the journey (`enterRegNumber` step) and found its only if it's missing a reg_uuid. The problem is `setup_registration()` in the registrations helper already checks for this and will throw an error if a registration is passed in with a missing reg_uuid. And `setup_registration()` is one of the first things called in `create()` so the subsequent call to `@registration.valid?` is pointless in this particular case.

This change removes the check to reduce the number of conditional branches in the `create()` method.

* Remove helper method formatIRRenewalNumber()

It was a method that simply strips then uppercases whatever string is passed in, and this can be done as `foo.strip.upcase`.

The method is therefore unnecessary, has no unit tests, and is currently only used in one place anyway.

This change removes it and updates the code that previously used it to do the work in place.

* Refactor existing IR logic in existing_reg_ctrlr

We will be adding new logic to handle new style registrations, which in turn will make the `ExistingRegistrationController` grow.

So prior to that this change refactors the code so the logic for determining if the entered registration number is a valid IR registration is moved to a new method.

This will help reduce complexity in the controller method and make its behaviour clearer.

* Refactor logic for digital regs in existing_reg_ctrlr

Inline with changes made for the handling of when an IR registration number is passed, this change updates how new style registration numbers are handled to follow the same pattern.

This will set us up for when we begin adding the actual functionality in subsequent PR's.