Add new find by reg. no. methid to Registration
None of the existing find methods allowed us to quickly and simply search for a registration by its reference.

We had to add a new feature to the waste-carriers-service project to handle the querying, so we have also added a new find method here that integrates with the query in services.