Update validation for if IR registration not found
Having updated the validation logic to distinguish between an invalid reg. number, and one what that does not match a current registration this change updates the logic around IR reg. numbers to ensure consistency. Now it will make a distinction between the format being invalid (reg not recognised) and from the java services when queried it returns no match (reg not found).