Logic to confirm if reg. exists in enter reg. page (#113)
https://eaflood.atlassian.net/browse/WC-84

As part of properly handling and directing users wishing to renew registrations created after March 2015 we need to confirm if the registration number entered actually exists.

This PR covers all changes related to implementing the logic to do this.

* Add new find by reg. no. method to Registration

None of the existing find methods allowed us to quickly and simply search for a registration by its reference.

We had to add a new feature to the waste-carriers-service project to handle the querying, so we have also added a new find method here that integrates with the query in services.

* Check if reg. exists in existing reg ctrllr

Now we have a way of querying the waste carrier services project to see if a registration exists using the registration number, we implement the check as part of the `ExistingRegistrationController.existing_registration?` method.

* Add validation handling if registration not found

Updated the validation logic in the `ExistingRegistrationController` when a user enters a new style registration number. Now it will make a distinction between the format being invalid (reg not recognised) and getting no result from the java services when queried it returns no match (reg not found).

* Update validation for if IR registration not found

Having updated the validation logic to distinguish between an invalid reg. number, and one what that does not match a current registration this change updates the logic around IR reg. numbers to ensure consistency. Now it will make a distinction between the format being invalid (reg not recognised) and from the java services when queried it returns no match (reg not found).