Add initial Irrenewal test and update factory
There was no spec for the `Irrenewal` model prior to this, and as we are about to add some logic to validate whether it is eligible for renewal we needed one.

As part of this also noted that the current factory didn't actually represent what an IR registration record looks like in the database. Each type has a different structure so this change updates the factory to be realistic.