Add in_renewal_window? to Irrenewal
The next thing we need to check before we can confirm an IR registration can be renewed is that it expires within the renewal window (currently 6 months in the service but due to be dropped to 3 months soon).

This change adds `in_renewal_window?` to `Irrenewal` which when called on an instance of `Irrenewal` will return true if the registration expires within the current renewal window.