Add already_renewed? to Irrenewal
Another check we want to do is see if the IR registration number has been used previously to renew a registration. This is to close a loophole where an IR registration number could be used multiple times to create new registrations at the lower renewal charge.

To support this change a new endpoint was added to the Java services layer to allow us to query registrations by their `originalRegistrationNumber` property.

You'll see from the unit tests we consider an IR renewal to have been used if its found on a regsistration whose status is not PENDING. In all other situations; the service being down or throwing an error, or its assigned to a registration that is not yet completed we return false.

This does mean there are potential edge cases where an IR renewal could be used more than once. However they are situations which are unlikely to happen and not worth the investment in time to pursue further options.