Exclude IRRenewal tests from being re-recorded
This change prevents the unit tests we added for `Irrenewal.already_renewed?` from being overwritten once created. This is because the tests rely on certain data existing in the database prior to running, to ensure the same cassettes would be created wherever the project is cloned and used.

However that is unlikely to be the case so having created the files, we always use them regardless of the global re-recording configiration rules.