Add VCR cassettes to the project
The project uses VCR to mock out http interactions between it and external services. Till now the cassettes would be recreated each time someone new cloned the project, however this generally led to time needing to be spent to either update the specs to match what the API was now actually returning, or to use keys that would return records that matched what was expected.

It wasn't actually about updating the specs because the API itself had changed, or the logic in the code. It was just the data being returned had changed and no longer matched what the test needed.

Therefore to stop this, this change brings back into the project the missing cassettes. It ensures our companies house auth key is not included with the files, and what's left is data captured when our internal services were returning dummy data.

Now the next dev that comes along and needs to run the rspec suite should find it passes on first pass.