Check if IR registration eligible to renew
https://eaflood.atlassian.net/browse/WC-84

Having refactored the functionality to validate the IR registration number entered by the user, and to confirm it exists, this change builds on these to confirm that the registration is eligible to renew.

It adds the logic to do the checks, and validation messages to display when the checks fail. The complete list of checks in the order they are applied are

- Check registration number is in a valid IR format
- Check IR registration exists (`irrenewals` in the db)
- Check registration has not expired
- Check the expiry date is in the renewal window for the registration
- Check the registration number has not been used to renew a another registration

If all these checks pass, the user is redirected to the business type page and allowed to continue with renewing their IR registration.