Add check and validation error for expired IR reg.
Added the check to see if the IR registration provided is expired, and if so display a suitable error to the user.

In doing so realised where the function was sat on the `Irrenewal` model was not suitable (we don't actually get one of those back) and didn't want to add it to `Registration` which is already overloaded.

So currently best fit feels like a private method on the controller itself. Hence movement of function and deletion of tests.