Refactor IR expired? check in existing_reg_ctrlr
Forseeing that `can_renew_ir_registration?()` would expand to quite a large method if I followed the previous pattern, decided instead to move the management of adding errors to the regisration object in the event of a check failing into the check itself. Now `can_renew_ir_registration?` is just a series of calls to these various checks.