Add handling 4 IR reg. not in renewal window
Added the check to see if the IR registration provided is eligble for renewal based on the current renewal window, and if not display a suitable error to the user.

As before the functionality for this sitting on the `Irrenewal` didn't make sense hence have also updated the class and its unit tests.