Check if IR registration eligible to renew (#114)
* Check if IR registration eligible to renew

https://eaflood.atlassian.net/browse/WC-84

Having refactored the functionality to validate the IR registration number entered by the user, and to confirm it exists, this change builds on these to confirm that the registration is eligible to renew.

It adds the logic to do the checks, and validation messages to display when the checks fail. The complete list of checks in the order they are applied are

- Check registration number is in a valid IR format
- Check IR registration exists (`irrenewals` in the db)
- Check registration has not expired
- Check the expiry date is in the renewal window for the registration
- Check the registration number has not been used to renew a another registration

If all these checks pass, the user is redirected to the business type page and allowed to continue with renewing their IR registration.

* Add initial Irrenewal test and update factory

There was no spec for the `Irrenewal` model prior to this, and as we are about to add some logic to validate whether it is eligible for renewal we needed one.

As part of this also noted that the current factory didn't actually represent what an IR registration record looks like in the database. Each type has a different structure so this change updates the factory to be realistic.

* Add expired? method to Irrenewal

This change adds a method that when called on an instance of `Irrenewal` will tell the caller whether the registration is expired or not.

This is the first check of eligibility as to whether the IR registration can be renewed.

* Add in_renewal_window? to Irrenewal

The next thing we need to check before we can confirm an IR registration can be renewed is that it expires within the renewal window (currently 6 months in the service but due to be dropped to 3 months soon).

This change adds `in_renewal_window?` to `Irrenewal` which when called on an instance of `Irrenewal` will return true if the registration expires within the current renewal window.

* Remove redundant it works test

Now we have other unit tests we don't need this anymore.

* Add already_renewed? to Irrenewal

Another check we want to do is see if the IR registration number has been used previously to renew a registration. This is to close a loophole where an IR registration number could be used multiple times to create new registrations at the lower renewal charge.

To support this change a new endpoint was added to the Java services layer to allow us to query registrations by their `originalRegistrationNumber` property.

You'll see from the unit tests we consider an IR renewal to have been used if its found on a regsistration whose status is not PENDING. In all other situations; the service being down or throwing an error, or its assigned to a registration that is not yet completed we return false.

This does mean there are potential edge cases where an IR renewal could be used more than once. However they are situations which are unlikely to happen and not worth the investment in time to pursue further options.

* Exclude IRRenewal tests from being re-recorded

This change prevents the unit tests we added for `Irrenewal.already_renewed?` from being overwritten once created. This is because the tests rely on certain data existing in the database prior to running, to ensure the same cassettes would be created wherever the project is cloned and used.

However that is unlikely to be the case so having created the files, we always use them regardless of the global re-recording configiration rules.

* Add check and validation error for expired IR reg.

Added the check to see if the IR registration provided is expired, and if so display a suitable error to the user.

In doing so realised where the function was sat on the `Irrenewal` model was not suitable (we don't actually get one of those back) and didn't want to add it to `Registration` which is already overloaded.

So currently best fit feels like a private method on the controller itself. Hence movement of function and deletion of tests.

* Refactor IR expired? check in existing_reg_ctrlr

Forseeing that `can_renew_ir_registration?()` would expand to quite a large method if I followed the previous pattern, decided instead to move the management of adding errors to the regisration object in the event of a check failing into the check itself. Now `can_renew_ir_registration?` is just a series of calls to these various checks.

* Add handling 4 IR reg. not in renewal window

Added the check to see if the IR registration provided is eligble for renewal based on the current renewal window, and if not display a suitable error to the user.

As before the functionality for this sitting on the `Irrenewal` didn't make sense hence have also updated the class and its unit tests.

* Add handling 4 IR reg. that is already renewed

Added the check to see if the IR registration provided has already been used to renew a previous registration.

As before the functionality for this sitting on the `Irrenewal` didn't make sense hence have also updated the class and its unit tests.

* Remove vcr cassettes that are no longer used

* Minor tweaks from review

Tweak how the "date users can renew from" is calculated. For example, if the current day is 12 February, users can only renew registrations where the expiry date is less than 3 months away (eg. 11 May). This was saying 3 months and 1 day way, but should be 3 months minus 1 day.

Also amend copy for an error message.