Add tests to clarify renewal window behaviour (#117)
These tests are mostly for documentation as there was a lot of confusion around how the dates worked (normally we'd want to avoid directly testing private methods). Users can renew from 3 months before the expiry date – so if it expires on 25 May, they can renew from 25 February to 24 May. This PR also corrects the date which users are told they can renew from.