Implement checks for digital registrations
This change adds a can renew method for digital registrations, and a new check that tests the status of the registraqtion being renewed. It must be ACTIVE to be eligible to be renewed.

It also makes changes to the existing code to allow this, namely ensuring the `@registration` object has the data subsequent methods will rely on.