Check if registration eligible to renew
https://eaflood.atlassian.net/browse/WC-84

Having refactored the functionality to validate the digital registration number entered by the user, and to confirm it exists, this change builds on these to confirm that the registration is eligible to renew.

It adds the logic to do the checks, and validation messages to display when the checks fail.

WIP