Check if registration eligible to renew (#118)
https://eaflood.atlassian.net/browse/WC-84

Having refactored the functionality to validate the digital registration number entered by the user, and to confirm it exists, this change builds on it to confirm that the registration is eligible to renew.

It adds the logic to do the checks, and validation messages to display when the checks fail. The checks are

- that it's not expired
- that it expires within the currently configured renewal window
- that its status is ACTIVE

In support of this we also added some unit tests for the private check methods, and the two `can_renew?` methods in `ExistingRegistrationController` as a form of documentation. They make it clear when a registration (whether digital or IR) can and cannot be renewed.