Add static page with info for users trying 2 renew
Chosen to redirect users to a static content page. If we tried to do anything dynamic to incorporate displaying the relevant registration number, we'd be required to perform the same checks again as the page is public and we don't want people going directly there, plus its only meant to be temporary. So we don't want the overhead additional logic and tests to cater for it.