Redirect digital renewals to temporary info page (#120)
https://eaflood.atlassian.net/browse/WC-84

Now we have built in robust checks that a registration is eligible for renewal we need to be in a position to redirect them to our renewals service when its live, and to a temporary page which explains the use of extensions whilst it isn't.

This change handles

- the ability to redirect to a configurable endpoint
- passing the registration number in a format the renewals service will recognise
- creating a new page which explains the use of extensions which users will be redirected to until renewals is live

For reference we have chosen to redirect users to a static content page because if we tried to do anything dynamic to incorporate displaying the relevant registration number, we'd be required to perform the same checks again. The page is public and we don't want people going directly there, plus its only meant to be temporary. So we don't want the overhead additional logic and tests to cater for it doing this would add.