Refactor expired? from ctrlr to service
This change refactors the core functionality of `ExistingRegistrationController.expired?` to a service object. This is because we'll need to call on the same function from within `Registration` as part of our work on displaying a renewal button in the registrations index page when a user is signed in.

This includes taking out the need to test a private method in the controller spec, and instead having a dedicated unit test in a new service spec.