Refactor in_renewal_window? into DateService
Completing the last refactor of a method that is used to determine if a registration can be renewed into something that can be reused cleanly elsewhere. This moves the date logic that determines if a registration is within the renewal window into the `DateService` ready for reuse in the registration itself.