Build into DateService conversion from millisecond
Even after our refactoring of the `ExistingRegistrationController` to move date dependent logic to the `DateService` we are still reliant on a method in the `ApplicationHelper` to handle converting the way dates are held against the registration (UTC milliseconds) to something a date we can use.

So this change adds logic to the `DateService` to handle this. Going forward we should just be able to instantiate a `DateService` object using something like

```
service = DateService.new(registration.expires_on)
```

and it will handle the conversion automatically, whilst still working if we instantiate it as `service = DateService.new(Date.today)`.