Start of new CanBeRenewed concern
Part of identifing whether a registration can be renewed is running a series of checks across it. We did these in `ExistingRegistrationController` but now need to do the same from the registration controller.

Stuff that can be reused we have moved out to the new `DateService` object but now there is stuff we need to add to the `Registration` object itself. Rather than add to the existing model which is already touching 2000 lines of code, we've decided instead to move this out to a concern.

This is the first step in that, in which we update and move the `Registration.expired?` method to the new `CanBeRenewed` concern (plus add some tests).