Refactor Registration concern tests to shared
Rather than overloading the registration model spec itself with tests for things in a concern, implemented a strategy for putting concern tests in a shared examples spec and then adding a single `it_behaves_like` in the model spec.

This was guided by the following examples

- https://stackoverflow.com/a/20010923/6117745
- https://nicholaide.github.io/rspec/2016/10/29/rspec-test-concerns.html
- https://modocache.io/shared-examples-in-rspec

There doesn't appear to be any concensus on where to put them, so I chose to place them in `spec/model/concerns` to mirror where they sit in the project structure.