Add helper to get date in full as a string
Add a new helper that takes a date and returns the day, date as a ordinal, month and year as a string. So for example 2018-02-21 would be returned as "Wednesday 21st February 2018".

We need dates in this format for the validation message when a user attempts to renew a registration that is outside the renewal window, and the code was copied from elsewhere so moving it here hopefully means we stop the copy and pasting and start reusing 😊 !