Add ability to log val. errors in can_renew?
We need to check if a registration is eligible for renewal in the `ExistingRegistrationController` and in `Registration` so we know whether to display a renewal link. So it makes sense to use the same code in both places.

However in one of those we also need to be able to log a validation error message depending on why the registration cannot be renewed (`ExistingRegistrationController`). So to support the same checks being used for both, this change adds the ability to tell `CanBeRenewed.can_renew?` to also log error messages (the default is it doesn't).