Fix broken tests for in_renewal_window?
The tests for `DateService.in_renewal_window?` pass in a `Time` object and having run `date_service_spec.rb` fully it highlighted we got an error in `date_from_utc_milliseconds` because it cannot handle an instance of `Time` being passed in.

So this change covers this edge case and fixes it.