Add link to renew reg. from user signed in page (#121)
https://eaflood.atlassian.net/browse/WC-213

Currently the only way for users to renew a registration is via the external site and the start a new registration journey. At the start they are asked if they actually want to renew, and from there they will be able to access the new renewals service.

This adds another route to renew via the page users see once they have signed in. When signed in users see a list of registrations. This change adds

- Determining for each registration displayed whether it is eligible to renew
- Against each eligible registration adding a button that allows them to access the renewals service
- Adding a renew link to each registration in the back office results screen for each reg. that is eligible

To do this we needed to refactor the checks that were being done in the `ExistingRegistrationController` into a reusable form; still used by the controller but also by the `Registration` object. There were also some other changes made to support this work.

* Refactor expiry date checks to service

The checks we were doing on an expiry date (is it expired, is it in the renewal window) are now done in a new `ExpiryDateService` service object. It also handles taking the date as is in the registration (String representation of milliseconds since the UTC epoch) and converting it into a format we can then run our logic on.

* Refactor `can_renew?` to a concern

Rather than call a method in the `ExistingRegistrationController` to determine whether a registration can be renewed, we now ask it! This is achieved by adding a new concern (this projects first) and including it in the `Registration` object.

A thing to note is that rather than overloading the registration model spec itself with tests for things in a concern, we implemented a strategy for putting concern tests in a shared examples spec and then adding a single `it_behaves_like` in the model spec.

This was guided by the following examples

- https://stackoverflow.com/a/20010923/6117745
- https://nicholaide.github.io/rspec/2016/10/29/rspec-test-concerns.html
- https://modocache.io/shared-examples-in-rspec

There doesn't appear to be any consensus on where to put them, so we chose to place them in `spec/model/concerns` to mirror where they sit in the project structure.

* Fix lower tier regs being eligible for renewal

The previous checks were not incorporating lower tier registrations. Now not only do we ensure they are not redirected to the renewals service, we display a useful validation message if a user enters one in the 'renew a registration' screen.

* Add a date formatter

When a user enters a registration number whose expiry date is not in the renewal window we display a validation message that includes the date it can be renewed from. We added a new Date formatter to support this

* Fix issue caused by renaming of method

This change fixes an error flagged in testing, caused by a call to a method that has since been renamed.

It also removes unnecessary checks on whether the value exists before the calling `valid_ir_format?` as this is handled in the method.