Refactor static page to a controller and view
Because we will be introducing dynamic elements to the page we display based on the registration being renewed, it no longer counts as a static page.

So this change moves the logic out of the static pages controller and into a new one specifically for handling and displaying the renewal information page.