Update content CBD renewals service stand in page
https://eaflood.atlassian.net/browse/WC-84

Until the renewals service goes live anyone entering a CDB registration number that is eligible for renewal will be redirected to a 'stand in' page. This is intended to explain the current situation and advice users what actions they need to take.

The initial version was based on the regulatory position statement (RPS) however our colleagues in NCCC felt this was to vague in areas and would rather it more closely mirrored the content of the renewals letters they are sending out.

So this change is intended to bring the content of the page in line with that need.

WIP