Remove external link icon from link in page to RPS
Had it in my head that it had become the norm to include the external link icon in links that take you out of the service.

However a discussion at work led me to question this and finally I tracked down the following post [Removing the external link icon from GOV.UK](https://designnotes.blog.gov.uk/2016/11/28/removing-the-external-link-icon-from-gov-uk/).

So to ensure this matches what is being done in other work, and to follow current guidance from GDS this change removes the element in the html which causes the icon to display.