Update content CBD renewals service stand in page (#123)
* Update content CBD renewals service stand in page

https://eaflood.atlassian.net/browse/WC-84

Until the renewals service goes live anyone entering a CDB registration number that is eligible for renewal will be redirected to a 'stand in' page. This is intended to explain the current situation and advice users what actions they need to take.

The initial version was based on the regulatory position statement (RPS) however our colleagues in NCCC felt this was to vague in areas and would rather it more closely mirrored the content of the renewals letters they are sending out.

So this change is intended to bring the content of the page in line with that need.

We also updated the external link to the RPS to remove the icon.

Initially believed it was the norm to include the external link icon in links that take you out of the service.

However a discussion at work led me to question this and finally I tracked down the following post [Removing the external link icon from GOV.UK](https://designnotes.blog.gov.uk/2016/11/28/removing-the-external-link-icon-from-gov-uk/).

So to ensure this matches what is being done in other work, and to follow current guidance from GDS this change also removes the element in the html which causes the icon to display.