Fix or mark PENDING broken feature tests
Those tests that have been marked PENDING have so because our connection to the test Worldpay environment is currently broken. Due to changes they have made to enhance security (changes expected to come to their production service in May) we are now unable to connect to Worldpay. The fix involves whitelisting a parameter in the test admin interface, something only the EA finance team can do. So till then all we can do is mark any test that makes a credit card payment as pending.

The rest of the changes are due to the use of values that are either no longer present (for example the results returned by address lookup) or no longer valid (use of inactive company registration numbers).